#10 - Code refactor and cleanup
