+++ src/plugins/sidebar-comments/containers/Sidebar.js
@@ -14,7 +14,7 @@ import { withSelect } from '@wordpress/data';
 /**
  * Plugin Imports.
  */
-import withPostMeta from '../../../components/higher-order/withPostMeta';
+import withPostMeta from '../../containers/higher-order/withPostMeta';
 // Import the component that we are going to wrap.
 import Sidebar from '../components/Sidebar';
